add preliminary global styles in the css file
To make the CSS file easier to read and manage, it is has been
organized based on which webpage the style applies to.

Global styles include the title and navigation bar that will
be present on ALL webpages.

This styling is considered preliminary because there is no color
or theming applied! This will be added at a later date and
was done intentionally to make styling the webpage more
efficient, concise, and less overwhelming